mail_retrieval/pyproject.toml
[project]
name = "mail-retrieval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-pinecone>=0.2.12",
    "python-dotenv>=1.1.1",
]
